+++ app/layout.tsx
@@ -4,10 +4,10 @@ import Image from "next/image";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 import { Analytics } from "@vercel/analytics/next";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import NavBar from "@/components/NavBar";
-import { SearchDialogProvider } from "@/components/SearchDialogProvider";
-import QueryProvider from "@/components/QueryProvider";
-import { LoginDialogProvider } from "@/components/LoginDialogProvider";
+import NavBar from "@/components/navbar/NavBar";
+import { SearchDialogProvider } from "@/components/dialogs/SearchDialogProvider";
+import QueryProvider from "@/components/providers/QueryProvider";
+import { LoginDialogProvider } from "@/components/auth/LoginDialogProvider";
 import "@/app/globals.css";
 
 const hahmlet = Hahmlet({
