+++ components/auth/CurrentUserAvatar.tsx
