+++ components/jargon/JargonCard.tsx
